bcast/file_bcast - reinit static local variables on each new file.
The _get_block_[none|lz4] family of functions declare some local static
variables to keep track and maintain values of stuff like the remaining
bytes or the file position across file blocks.

If we are getting the first block of a new file, these values need to be
reinitialized.

Bug 11427.